+++ packages/ui-service-inquirer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/ui-service-inquirer
 
+## 0.2.50
+
+### Patch Changes
+
+-   Updated dependencies [caff0ca]
+    -   @sap-ux/odata-service-inquirer@2.11.5
+
 ## 0.2.49
 
 ### Patch Changes